requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements.in'
#
coloredlogs==15.0.1
    # via
    #   onnxruntime
    #   onnxruntime-qnn
filelock==3.16.1
    # via torch
flatbuffers==24.3.25
    # via
    #   onnxruntime
    #   onnxruntime-qnn
fsspec==2024.10.0
    # via torch
humanfriendly==10.0
    # via coloredlogs
jinja2==3.1.4
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==1.24.4
    # via
    #   -r .\requirements.in
    #   onnx
    #   onnxruntime
    #   onnxruntime-qnn
    #   opencv-python
    #   torchvision
onnx==1.17.0
    # via -r .\requirements.in
onnxruntime==1.20.0
    # via -r .\requirements.in
onnxruntime-qnn==1.20.0
    # via -r .\requirements.in
opencv-python==4.10.0.84
    # via -r .\requirements.in
packaging==24.1
    # via
    #   onnxruntime
    #   onnxruntime-qnn
pillow==11.0.0
    # via
    #   -r .\requirements.in
    #   torchvision
protobuf==5.28.3
    # via
    #   onnx
    #   onnxruntime
    #   onnxruntime-qnn
pyreadline3==3.5.4
    # via humanfriendly
sympy==1.13.1
    # via
    #   onnxruntime
    #   onnxruntime-qnn
    #   torch
torch==2.5.1
    # via
    #   -r .\requirements.in
    #   torchvision
torchvision==0.20.1
    # via -r .\requirements.in
typing-extensions==4.12.2
    # via torch
